Implementation of handlers with dynamic signatures
Now it's possible to have a handler that receives a set of parameters
that's not simply `(request)`.

You can write your handlers to receive your own models, for this all you
have to do is to populate `request["types_registry"]` with all types
that you want to make available to handlers.